made float

+++ afsk/main.c
@@ -923,7 +923,7 @@ int get_tlm_fox() {
 	digitalWrite (0, LOW);
 	printf("Sending %d buffer bytes over socket!\n", ctr);
 //	int sock_ret = send(sock, buffer, buffSize, 0);
-	int sock_ret = send(sock, buffer, (int)(ctr * ( ((int)(frameCnt/2 + 0.5))/frameCnt )), 0);
+	int sock_ret = send(sock, buffer, (int)(ctr * ((float)( ((int)(frameCnt/2 + 0.5)))/(float)(frameCnt) )), 0);
 	printf("Result of socket send: %d \n", sock_ret);
 	if (sock_ret == -1)  {
 		printf("Error: %s \n", strerror(errno));
